Default to using Host header from request as ExternalURL
Instead of defaulting to the bound ip and port when `-external-url` is not given,
use the value of the Host header in the request requesting a resumable upload.

This allows you to refer to the server by any name and it will have you use the same one again
when doing an upload.

Since `Server.URL()` is also used in tests, allow passing in `nil` to get the old behaviour.